requirements.txt
aiohttp==3.8.1
aiogram==2.21
python-dotenv==0.21.0
